Bellman {watervalues}	R Documentation

Compute Bellman values at step i from step i+1

Description

Compute Bellman values at step i from step i+1

Usage

Bellman(
  Data_week,
  next_week_values_l,
  decision_space,
  E_max,
  P_max = 0,
  niveau_max,
  method,
  na_rm = TRUE,
  max_mcyear,
  print_test = FALSE,
  correct_outliers = FALSE,
  q_ratio = 0.75,
  test_week,
  counter,
  inaccessible_states = 1,
  stop_rate = 5,
  ...
)

Arguments

Data_week
A "data.table" generated in Grid_Matrix code that contains: * states Numeric. All the water values that can be set, listed in decreasing order. * value_inflow Numeric. Inflow values for each Monte-Carlo year. * Rewards for each simulation value and each Monte-Carlo year. * level_high Numeric. Highest possible reservoir value. * level_low Numeric. Lowest possible reservoir value. * states_next List of vectors enumerating all reachable states

next_week_values_l
Numeric. Bellman values at step i+1.

decision_space
Simulation constraints values

E_max
Numeric of length 1. Maximum energy that can be generated by hydro storage over one step of time.

P_max
Numeric of length 1. Maximum energy that can be pumped to reservoir over one step of time.

niveau_max
Numeric of length 1. Reservoir capacity.

method
Character. Perform mean of grids algorithm or grid of means algorithm or grid of quantile algorithm.

na_rm
Boolean. Remove NAs

max_mcyear
Numeric of length 1. Number of Monte-Carlo year used in simulations

print_test
Boolean. print Bellman values.

correct_outliers
If TRUE, outliers in Bellman values are replaced by spline interpolations. Defaults to FALSE.

q_ratio
numeric in [0,1]. the probability used in quantile algorithm.

test_week
Numeric of length 1. number of the week to print in test.

counter
Numeric of length 1. number of the week in calculation.

inaccessible_states
Numeric in [0,1]. Tolerance of inaccessible states. For example if equal to 0.9 we delete the state if this states is inaccessible by 90% of scenarios.

stop_rate
the percent from which the calculation stop. for example stop_rate=5 means the calculation stop if there is a week with less then 5% accessibles states.

...
further arguments passed to or from other methods.

Value

a data.table like Data_week with the Bellman values

[Package watervalues version 0.2.0 Index]