Grid_Matrix {watervalues}	R Documentation

Calculate grid layer matrix of Bellman values and water values

Description

Calculate grid layer matrix of Bellman values and water values

Usage

Grid_Matrix(
  area,
  simulation_names,
  reward_db = NULL,
  inflow = NULL,
  simulation_values = NULL,
  nb_cycle = 1L,
  district_name = "water values district",
  mcyears = NULL,
  week_53 = 0,
  states_step_ratio = 0.01,
  reservoir_capacity = NULL,
  na_rm = FALSE,
  correct_outliers = FALSE,
  method,
  only_input = FALSE,
  q_ratio = 0.5,
  monotonic_bellman = FALSE,
  test_week = NULL,
  opts = antaresRead::simOptions(),
  shiny = F,
  inaccessible_states = 1,
  until_convergence = F,
  convergence_rate = 0.9,
  convergence_criteria = 1,
  cycle_limit = 10,
  pumping = F,
  efficiency = 1,
  stop_rate = 5,
  ...
)

Arguments

area
An 'antares' area.

simulation_names
Names of simulations to retrieve.

reward_db
a table contains the reward values generated by the function get_Reward If it's NULL it auto calculated.

inflow
a table contains the hydro storage generated by the function readAntares with th option hydrostorage = TRUE. If it's NULL it auto calculated.

simulation_values
Values for the simulation.

nb_cycle
Number of times to run the algorithm.

district_name
Name of the district used to store output.

mcyears
MC years to consider, by default all of them.

week_53
Water values for week 53, by default 0.

states_step_ratio
Discretization ratio to generate steps levels between the reservoir capacity and zero . Defaults to 0.05

reservoir_capacity
Reservoir capacity for the given area in MWh, if NULL (the default), value in Antares is used if available else a prompt ask the user the value to be used.

na_rm
Remove NAs

correct_outliers
If TRUE, outliers in Bellman values are replaced by spline interpolations. Defaults to FALSE.

method
Perform mean of grids algorithm or grid of means algorithm or grid of quantile algorithm.

only_input
if TRUE skip bellman values calculation and return the input

q_ratio
from 0 to 1. the probability used in quantile method to determine a bellman value which q_ratio all bellman values are equal or less to it. (quantile(q_ratio))

monotonic_bellman
force increasing bellman values with the stock level in the calculation.

test_week
the week number u want to see it's calculation information in the console

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

shiny
Boolean. True to run the script in shiny mode.

inaccessible_states
Numeric in [0,1]. Tolerance of inaccessible states. For example if equal to 0.9 we delete the state if this states is inaccessible by 90% of scenarios.

until_convergence
Boolean. TRUE to repeat cycle until convergence or attending the limit.

convergence_rate
from 0 to 1. Define the convergence level from which we suppose that no need to continue another cycle..

convergence_criteria
the value define convergence. if the difference between two water values is less then this value those values are converged.

cycle_limit
Define cycles limit when you are in the until_convergence mod.

pumping
Boolean. True to take into account the pumping.

efficiency
in [0,1]. efficient ratio of pumping.

stop_rate
the percent from which the calculation stop. for example stop_rate=5 means the calculation stop if there is a week with less then 5% accessibles states.

...
further arguments passed to or from other methods.

Value

a data.table

[Package watervalues version 0.2.0 Index]