bellman_parallel_value {watervalues}	R Documentation

Compute Bellman values at step i from step i+1 (version used in parallel computing)

Description

Compute Bellman values at step i from step i+1 (version used in parallel computing)

Usage

bellman_parallel_value(
  Data_week,
  i,
  next_week_values_l,
  max_mcyear,
  E_max,
  P_max,
  niveau_max,
  na_rm,
  decision_space,
  alpha,
  decimals
)

Arguments

Data_week
A "data.table" generated in Grid_Matrix code that contains: * states Numeric. All the water values that can be set, listed in decreasing order. * value_inflow Numeric. Inflow values for each Monte-Carlo year. * Rewards for each simulation value and each Monte-Carlo year. * level_high Numeric. Highest possible reservoir value. * level_low Numeric. Lowest possible reservoir value. * states_next List of vectors enumerating all reachable states

i
the row number of the point to calculate her bellman value.

next_week_values_l
Numeric. Bellman values at step i+1.

max_mcyear
Numeric of length 1. Number of Monte-Carlo year used in simulations

E_max
Numeric of length 1. Maximum energy that can be generated by hydro storage over one step of time.

P_max
Numeric of length 1. Maximum energy that can be pumped to reservoir over one step of time.

niveau_max
Numeric of length 1. Reservoir capacity.

na_rm
Boolean. Remove NAs

decision_space
Simulation constraints values

alpha
maximum tolerance.

decimals
number of decimals.

Value

a data.table like Data_week with the Bellman values

[Package watervalues version 0.2.0 Index]