check_Bellman_inc {watervalues}	R Documentation

Check the monotonicity of the calculated Bellman values and return a success rate.

Description

Check the monotonicity of the calculated Bellman values and return a success rate.

Usage

check_Bellman_inc(results)

Arguments

results
A data.table contains the Bellman values Obtained using the function Grid_Matrix()

[Package watervalues version 0.2.0 Index]