check_resh_vu_dec {watervalues}	R Documentation

Check the monotonicity of the calculated the water values after putting them in Antares Format and return a success rate.

Description

Check the monotonicity of the calculated the water values after putting them in Antares Format and return a success rate.

Usage

check_resh_vu_dec(reshaped_results)

Arguments

reshaped_results
A matrix contains the water values Obtained using the function to_Antares_Format()

[Package watervalues version 0.2.0 Index]