constraint_generator {watervalues}	R Documentation

Generate the list of constraint values of the link between the fictive area and the real one

Description

Generate the list of constraint values of the link between the fictive area and the real one

Usage

constraint_generator(
  area,
  nb_disc_stock,
  pumping = F,
  pumping_efficiency = NULL,
  opts
)

Arguments

area
The area concerned by the simulation.

nb_disc_stock
Number of simulation to launch, a vector of energy constraint.

pumping
Boolean. True to take into account the pumping.

pumping_efficiency
between 0 and 1. the pumping efficiency ratio.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

[Package watervalues version 0.2.0 Index]