getPumpEfficiency {watervalues}	R Documentation

Get the Pumping efficiency ratio for an area reservoir

Description

Get the Pumping efficiency ratio for an area reservoir

Usage

getPumpEfficiency(area, force = FALSE, opts = antaresRead::simOptions())

Arguments

area
An 'antares' area.

force
If "reservoir management" is disabled, return anyway the reservoir capacity?

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Value

the reservoir capacity (in MWh), or NULL if none.

Examples

## Not run: 

getReservoirCapacity("fr")

## End(Not run)

[Package watervalues version 0.2.0 Index]