get_Reward {watervalues}	R Documentation

Get the reward matrix from simulations

Description

Get the reward matrix from simulations

Usage

get_Reward(
  simulation_res = NULL,
  simulation_names = NULL,
  pattern = NULL,
  district_name = "water values district",
  opts = antaresRead::simOptions()
)

Arguments

simulation_res
list of the simulation names and constraint values to obtain the reward.

simulation_names
generated by runWaterValuesSimulation

pattern
A pattern to identify simulations.

district_name
Name of the district used to store output.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Value

a data.table timeid,MCyear,simulation overall cost

[Package watervalues version 0.2.0 Index]