get_max_hydro {watervalues}	R Documentation

Get max hydro power that can be generated in a week

Description

Get max hydro power that can be generated in a week

Usage

get_max_hydro(area, opts = antaresRead::simOptions())

Arguments

area
The area concerned by the simulation.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

[Package watervalues version 0.2.0 Index]