mean_or_inf {watervalues}	R Documentation

Calculate the mean if there is no infinite or missing value. Return -Inf in the other case.

Description

Calculate the mean if there is no infinite or missing value. Return -Inf in the other case.

Usage

mean_or_inf(x, inaccessible_states)

Arguments

x
numeric vector whose mean is wanted.

inaccessible_states
Numeric in [0,1]. Tolerance of inaccessible states. For example if equal to 0.9 we delete the state if this states is inaccessible by 90% of scenarios.

[Package watervalues version 0.2.0 Index]