next_week_value {watervalues}	R Documentation

An intermediate function in bellman values calculation. Select the next week bellman values of a position from all bellman values.

Description

An intermediate function in bellman values calculation. Select the next week bellman values of a position from all bellman values.

Usage

next_week_value(next_week_values_l, max_mcyear, i)

Arguments

next_week_values_l
list of all bellman values of the next week.

max_mcyear
the maximum number of scenarios.

i
the number of row in the table of calculation.

[Package watervalues version 0.2.0 Index]