plot_Bellman {watervalues}	R Documentation

Plot Bellman and Water values

Description

Plot Bellman and Water values

Usage

plot_Bellman(
  value_nodes_dt,
  week_number,
  param = "vu",
  states_step_ratio = 0.01,
  bellman_week = NULL,
  ...
)

Arguments

value_nodes_dt
A data.table contains the Bellman and water values . Obtained using the function Grid_Matrix()

week_number
Numeric of length 1. number of the week to plot.

param
string contains the element to plot * "vu" to plot only water values * "b" to plot only bellman values * "both" to plot both water and bellman values Default "vu"

states_step_ratio
put the ratio to change reservoir discretization in percent 0.01 to augment by 1%

bellman_week
Numeric of length 1. number of the week to plot the correspondent Bellman values.

...
further arguments passed to or from other methods.

Value

a ggplot object

[Package watervalues version 0.2.0 Index]