plot_generation {watervalues}	R Documentation

Plot Turbining power Graph and return result table

Description

Plot Turbining power Graph and return result table

Usage

plot_generation(
  area,
  timestep = "daily",
  Mcyear = NULL,
  min_path,
  max_path,
  simulation_name = NULL,
  opts = antaresRead::simOptions()
)

Arguments

area
An 'antares' area.

timestep
Resolution of the data to import: weekly (default, a linear interpolation is done on the data), monthly (original data).

Mcyear
precise the MC year to plot. Null plot the synthesis. Default NULL

min_path
path of Pmin file "/user/Pmin **.txt"

max_path
path of Pmax file "/user/Pmax **.txt"

simulation_name
simulation name to plot.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Value

a ggplot object

[Package watervalues version 0.2.0 Index]