plot_reservoir {watervalues}	R Documentation

Plot Reservoir Guide Graph and return result table

Description

Plot Reservoir Guide Graph and return result table

Usage

plot_reservoir(
  area,
  timeStep = "weekly",
  mcyear = NULL,
  simulation_name = NULL,
  opts = antaresRead::simOptions(),
  shiny = F,
  only_g = F
)

Arguments

area
An 'antares' area.

timeStep
Resolution of the data to import: weekly (default, a linear interpolation is done on the data), monthly (original data).

mcyear
precise the MC year to plot. all to plot all years. Null plot the synthesis. Default NULL

simulation_name
simulation name to plot.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

shiny
Boolean. True to run the script in shiny mod.

only_g
Boolean. True plot only constraints graph.

Value

a ggplot object

[Package watervalues version 0.2.0 Index]