plot_reward_variation {watervalues}	R Documentation

Plot the reward variation and return the results in table

Description

Plot the reward variation and return the results in table

Usage

plot_reward_variation(
  reward_base,
  week_id,
  constraints_values = NULL,
  output = FALSE
)

Arguments

reward_base
A data.table contains the rewards. Obtained using the function get_Reward()

week_id
Numeric of length 1. number of the week to plot.

constraints_values
the value of the constraint

output
Boolean. TRUE to return the table of values with the plot.

Value

a ggplot object

[Package watervalues version 0.2.0 Index]