readReservoirLevels {watervalues}	R Documentation

Read Reservoir Levels

Description

Read Reservoir Levels

Usage

readReservoirLevels(
  area,
  timeStep = "weekly",
  byReservoirCapacity = TRUE,
  opts = antaresRead::simOptions()
)

Arguments

area
An 'antares' area.

timeStep
Resolution of the data to import: weekly (default, a linear interpolation is done on the data), monthly (original data).

byReservoirCapacity
Multiply the result by the reservoir capacity (if available).

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Value

a data.table

[Package watervalues version 0.2.0 Index]