resetPumpPower {watervalues}	R Documentation

Reset to 0 the pumping power

Description

Reset to 0 the pumping power

Usage

resetPumpPower(area, path = NULL, opts = antaresRead::simOptions())

Arguments

area
A valid Antares area.

path
Optional, a path where to save the hydro storage file.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Value

An updated list containing various information about the simulation.

Note

The function makes a copy of the original hydro storage time series, you can restore these with restoreHydroStorage.

See Also

restoreHydroStorage

[Package watervalues version 0.2.0 Index]