restorePumpPower {watervalues}	R Documentation

Restore the Pumping power series

Description

Restore the Pumping power series

Usage

restorePumpPower(
  area,
  path = NULL,
  opts = antaresRead::simOptions(),
  silent = F
)

Arguments

area
A valid Antares area.

path
Path to a manual backup.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

silent
Boolean. True to run without messages.

Value

An updated list containing various information about the simulation.

[Package watervalues version 0.2.0 Index]