runWaterValuesSimulation {watervalues}	R Documentation

Run a simulation for calculating water values for a specific area

Description

Run a simulation for calculating water values for a specific area

Usage

runWaterValuesSimulation(
  area,
  simulation_name = "weekly_water_amount_%s",
  nb_disc_stock = 10,
  nb_mcyears = NULL,
  binding_constraint = "WeeklyWaterAmount",
  fictive_area = NULL,
  thermal_cluster = NULL,
  path_solver = NULL,
  wait = TRUE,
  show_output_on_console = FALSE,
  overwrite = FALSE,
  link_from = NULL,
  remove_areas = NULL,
  opts = antaresRead::simOptions(),
  shiny = F,
  otp_dest = NULL,
  file_name = NULL,
  pumping = F,
  launch_simulations = T,
  reset_hydro = T,
  ...
)

Arguments

area
The area concerned by the simulation.

simulation_name
The name of the simulation, s is a placeholder for the constraint value defined by nb_disc_stock.

nb_disc_stock
Number of simulation to launch, a vector of energy constraint will be created from maximum pumping power to the hydro storage maximum and of length this parameter.

nb_mcyears
Number of Monte Carlo years to simulate or a vector of years indexes to launch.

binding_constraint
Name of the binding constraint.

fictive_area
Name of the fictive area to create, argument passed to setupWaterValuesSimulation.

thermal_cluster
Name of the thermal cluster to create, argument passed to setupWaterValuesSimulation.

path_solver
Character containing the Antares Solver path, argument passed to runSimulation.

wait
Argument passed to runSimulation.

show_output_on_console
Argument passed to runSimulation.

overwrite
If area or cluster already exists, should they be overwritten?

link_from
area that will be linked to the created fictive area. If it's NULL it will takes the area concerned by the simulation.

remove_areas
Character vector of area(s) to remove from the created district.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

shiny
Boolean. True to run the script in shiny mod.

otp_dest
the path in which the script save Rdata file.

file_name
the Rdata file name.

pumping
Boolean. True to take into account the pumping.

launch_simulations
Boolean. True to to run the simulations.

reset_hydro
Boolean. True to reset hydro inflow to 0 before the simulation.

...
further arguments passed to or from other methods.

Note

This function have side effects on the Antares study used, a fictive area is created and a new district as well.

[Package watervalues version 0.2.0 Index]