simulation_names_values {watervalues}	R Documentation

simulations names and values generator

Description

simulations names and values generator

Usage

simulation_names_values(
  area,
  simulation_name,
  nb_disc_stock,
  pumping = F,
  opts = antaresRead::simOptions()
)

Arguments

area
The area concerned by the simulation.

simulation_name
The used name of the simulation.

nb_disc_stock
Number of simulation launched.

pumping
Boolean. True to take into account the pumping.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Note

This function have no side effects on the Antares study used. It's used in case you forget to save the output of script runWaterValuesSimulation to generate the output. The results are meaningless unless you already launched the runWaterValuesSimulation with the same parameters.

[Package watervalues version 0.2.0 Index]