to_Antares_Format {watervalues}	R Documentation

Convert water values to Antares format

Description

This function converts water values generated by meanGridLayer or waterValues to the format expected by Antares: a 365*101 matrix, where the rows are the 365 days of the year and the columns are round percentage values ranging from 0 to 100 assessing the reservoir level. Since meanGridLayer and waterValues output weekly values for an arbitrary number of reservoir levels, interpolation is performed on both scales in order to fit the desired format.

Usage

to_Antares_Format(data, constant = T)

Arguments

data
A 7-column data.table generated by watervalues::Grid_Matrix()

constant
Boolean. Generate daily constant values by week. FALSE to do interpolation.

Value

A 365*101 numeric matrix

[Package watervalues version 0.2.0 Index]