waterValuesViz {watervalues}	R Documentation

Visualize mean grid layer

Description

Visualize mean grid layer

Usage

waterValuesViz(Data, filtre_ratio = 1, show_negative = TRUE)

Arguments

Data
Output from Grid_Matrix.

filtre_ratio
in [1,0]. Define the percent to keep from water values eliminating the rest (extreme negatives and positives).

show_negative
Boolean. true to show negative water values and False to delete them from the graph.

Value

a ggplot object

[Package watervalues version 0.2.0 Index]